Fix plantsMgr assert by combining plant headers
